requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
exceptiongroup==1.1.1
greenlet==2.0.1
idna==3.4
iniconfig==2.0.0
jsonpath==0.82
packaging==23.1
playwright==1.33.0
pluggy==1.0.0
pyee==9.0.4
PyMySQL==1.1.0
pytest==7.3.1
pytest-base-url==2.0.0
pytest-dependency==0.5.1
pytest-playwright==0.3.3
python-slugify==8.0.1
PyYAML==6.0
requests==2.31.0
text-unidecode==1.3
tomli==2.0.1
typing-extensions==4.6.1
urllib3==2.0.2
Faker==19.2.0